LayoutTests/imported/w3c/web-platform-tests/svg/animations/animate-marker-orient-from-auto-to-auto-start-reverse-expected.txt


FAIL Animate SVGMarkerElement orientAttr from auto to auto-start-reverse assert_equals: expected (undefined) undefined but got (number) 0
